qrshilde/requirements.txt
qrcode
Pillow
python-dotenv
openai>=1.0.0
